move show_overlay to BaseView
